Extract ground truth generation into a library.
